Make depth scissor integer and use x, y, width, height semantics like in other parts of engine

+++ glprogs/stages/depth/depth.params.glsl
@@ -4,7 +4,7 @@ struct ShaderParams {
     mat4 modelViewMatrix;
 	mat4 textureMatrix;
 	vec4 color;
-	vec4 scissor;
+	uvec4 scissor;
 	vec2 alphaTest;
 	uvec2 texture;
 };
